add functions to talk with DRE
The functions are a mix of what is described on the ccylus website
and the ann_lwr reactor archetype developed by Teddy. Based on
the output databases, this archetype doesn't talk with the DRE
or produce power, as it should be. This commit
also changes the fuel commodity and recipe inputs to be vectors
so that a user can define multiple desired commodities and
recipes.