get back to working order
The main issue was not removing the object id in the get_material_trades
funciton. Things do not fully work as needed because having the
self.discharged logic correct breaks things when it runs. Also,
completely reversing everything breaks things, which at least
feels consistent. If everything is reversed, except the output
of self.discharge(), then it runs, though not correctly