remove recipe loop in get_material_bids
The loop for the recipes was creating a bid for each
possiblespent fuel recipe, which is unphysical. Each spent fuel
assembly will only have one composition. Instead, get the index of
the commodity name in fuel_outcommods and matches to the recipe
at the same index of fuel_outrecipes. Enforces the need for users to
have consistent ordering in their input file