change writing of spent comps,  calls to OpenMC
The call to OpenMC is only done once, not for each assembly. All of
the assemblies in the core are transmuted, to account for
multiple batches in the core. The spent fuel recipes are written to
a list of dictionaries that are fed directly to the transmute
method in Cyclus, instead of written to an xml file. Using this
method, the compositions are not updated in the database.