move spent fuel recording to discharge
If fuel is to be transmuted at each cycle end, then the
spent fuel resources don't need to be recording each time fuel
is transmuted. This also captures all of the fuel that is
dischraged from a reactor, not just the transmuted fuel. This only has
an impact when a facility is decommissioned during a simulation