update spent compositions and power levels
Previous power level of 100 MW and mass of 100 kg (arbitarily chosen)
were too high to correspond to a single pin cell model. Lower the power
level to match a burnup of 50 MWd/kgU, and the mass to
correspond to the volume and density of the fuel. Results still seem
a little funky, but it shoul dbe a more accurate/realstic simulation