change transmute call time, remove record supplyPower
Change the transmute call to only transmute one assembly
at the time of discharge for the entire operating time
(all of the cycles at once). This is only a temporary change,
to do a sanity check on the separated Pu mass when the whole core
is transmuted at the end of each cycle. I just wanted to keep
a record of this iteration of the archetype. The other change is
removing the lib.record_time_series("supplyPOWER") because this
caused conflicts when the Cycamore Reactor is also deployed in
a simulation, and I don't think this information is needed in
the database from OpenMCyclus